ActividadC/main/requirements.txt
streamlit>=1.35.0
opencv-python>=4.10.0
pytesseract>=0.3.10
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.9.0
